fixed some clang errors, maybe it stays this way

+++ muse_mcl_2d/include/muse_mcl_2d/samples/sample_clustering_2d.hpp
@@ -31,7 +31,7 @@ struct SampleClustering2D {
         indexation(indexation)
     {
         auto resolution = indexation.getResolution();
-        angular_bins = std::floor(2 * M_PI / resolution[2]);
+        angular_bins = static_cast<int>(std::floor(2 * M_PI / resolution[2]));
     }
 
     inline void clear()
